Make path optimization more coarse.
The approximation is now 2.0/dpi instead of 1.1/dpi. This reduces
the number of path segments by about 40% and avoids reliably
single-pixel stair steps.

To get the same number of segments as before, the dpi value can
be raised. This gives even more overall accuracy, as the pixels
them selfs are more accurate.